[DAS-5470] Flag-control and disable-by-default glue table validation
*What this PR does / why we need it:*
Glue checks for five table params to indicate they're valid hive tables:
TableType
SerdeLib/ SerializationLib
InputFormat
OutputFormat
SerdeInfo

This is not in line with what okera needs. We're putting this logic behind a flag
and disablign it by default - thereby depending on okera's validation logic for
table metadata.

Additionally fixing an NPE in checking for presence of an environment variable.

*Special notes for your reviewer:*

*Additional testing that was done:*

```
Release note:
```